+++ stream/mock_publisher.go
@@ -10,11 +10,11 @@ type MockPublisher struct {
 	mock.Mock
 }
 
-func (m *MockPublisher) PublishLogs(strm string, id *int64, phase string, logs string) {
+func (m *MockPublisher) PublishLogs(strm string, id *int, phase string, logs string) {
 	m.MethodCalled("PublishLogs", strm, id, phase, logs)
 }
 
-func (m *MockPublisher) PublishObject(strm string, id *int64, event EventType, object interface{}) {
+func (m *MockPublisher) PublishObject(strm string, id *int, event EventType, object interface{}) {
 	m.MethodCalled("PublishObject", strm, id, event, object)
 }
